requirements.txt
streamlit
requests
beautifulsoup4
pandas
firecrawl-py
tqdm
